add jwt and remove some lombok from entity classes

+++ Wamya-backend-application/src/main/java/com/excentria_it/wamya/domain/JourneyRequestSearchDto.java
@@ -17,7 +17,7 @@ public interface JourneyRequestSearchDto {
 	@Value("#{T(com.excentria_it.wamya.application.utils.MapperUtility).buildEngineTypeDto(target.engineTypeId, target.engineTypeName)}")
 	EngineTypeDto getEngineType();
 
-	Integer getDistance();
+	Double getDistance();
 
 	LocalDateTime getDateTime();
